refactor(foldkit): move the perpetual fibers out of the boot closure
`startWith` forked the Subscription streams and the Managed Resource
lifecycles inline: one fiber per Subscription that derives its
dependencies from every Model, restarts its stream when they change,
and feeds its Messages into the queue, and one fiber per Managed
Resource that derives its requirements the same way, acquires and
releases the resource as they change, and writes the shared Ref.

`forkSubscriptionFibers` in `subscriptionFibers.ts` and
`forkManagedResourceFibers` in `managedResourceFibers.ts` now do that.
Each is a plain function returning the Effect that forks the fibers
into the runtime scope, and the boot closure yields the two calls where
the inline code stood, handing them the init Model, the Model PubSub,
the runtime scope, the queue's `enqueueMessageEffect`, and `crashWith`.
The Subscription unit also takes the slow-phase config for dependency
derivation and `provideAllResources`. The Managed Resource unit takes
the refs the resource provider made. The moved code is unchanged apart
from formatting and the two `yield*` expressions becoming the
functions' return values.

`ResolvedSlowPhaseConfig` is now exported from `slowPhase.ts`, because
the Subscription unit names it in its parameter type. The NOTE in the
Managed Resource fork lost its opening clause about mirroring the
Subscription fork, which now lives in another file, and keeps why a
defect crashes the runtime and why the fork needs no resources.

No user-facing effect, so the changeset is empty.
